solve casting bug

+++ aula-01/java/src/main/java/com/github/newtonjose/ufg/cs/aula01/AlgoritmosUtils.java
@@ -368,8 +368,8 @@ public final class AlgoritmosUtils {
 
         final int aux = num % ConstAuxUtils.DIVISOR_CEM;
 
-        final double centena = num / ConstAuxUtils.DIVISOR_CEM;
-        final double dezena = aux / ConstAuxUtils.DIVISOR_DEZ;
+        final int centena = (int) (num / ConstAuxUtils.DIVISOR_CEM);
+        final int dezena = (int) (aux / ConstAuxUtils.DIVISOR_DEZ);
 
         final int prop = (int) (Math.pow(centena, ConstAuxUtils.NUM_TRES)
                         + Math.pow(dezena, ConstAuxUtils.NUM_TRES)
